backend/requirements.txt
# Pipecat AI with required service integrations
pipecat-ai[assemblyai,openrouter,elevenlabs,mem0]

# Environment variable management
python-dotenv

# Enhanced logging
loguru 
pyyaml

# Local control plane (HTTP + WebSocket)
fastapi
uvicorn[standard]
 
# Twitch IRC client
twitchio